STAT 301: Final Project
Through this STAT 301 project, we attempt to build a model to analyze student performance based on many factors such as their study time, relationship status and absences etc.

This dataset was obtained from the UCI Machine Learning Repository and it was collected through school reports and questionnaires.

The statistical question that we’ll be answering through this project is whether certain factors such as internet access, study time, relationship status, going out with friends and their absences have an impact on whether or not a student fails a class. Our model is based on inference - understanding what factors lead to a student failing in class.